requirements.txt
discord.py
easyocr
pillow
numpy
python-dotenv